+++ liblick/llist.h
@@ -29,7 +29,7 @@ node_t *new_node(void *val, node_t *next);
  * @param lst the list to count
  * @return the length of the list
  */
-size_t list_length(node_t *lst);
+int list_length(node_t *lst);
 /**
  * @brief filter a list into two separate lists using a predicate
  *
